Gen-DeepResearch-RL/rllm/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rllm"
version = "0.2.1"
description = "Reinforcement Learning for Language Agents"
authors = [
    {name = "rLLM Team"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
license = {file = "LICENSE"}
dependencies = [
    # Core ML/AI packages
    "torch",
    "torchvision",
    "transformers>=4.55.0",

    # Data processing
    "datasets>=2.20.0",
    "pandas",
    "polars",
    "pillow",

    # Infrastructure
    "ray",

    # Math and science
    "numpy",
    "sympy",
    "pylatexenc",
    "antlr4-python3-runtime==4.9.3",

    # Tools
    "mcp",

    # Integration with Eval Protocol
    "eval-protocol",

    # Utilities
    "hydra-core",
    "openai",
    "fastapi",
    "uvicorn",
    "tqdm",
    "PyYAML",
    "pydantic",
    "wrapt",
    "asgiref>=3.7.0",
    "wandb",
    "json5",
    "aiosqlite",
    "oss2",
    "requests==2.32.3",
]

[project.optional-dependencies]

dev = [
    "pytest",
    "pre-commit",
    "ruff",
    "mypy",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-autorefs>=0.5.0",
    "pymdown-extensions>=10.0.0",
]

verl = [
    "verl==0.6.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    # "vllm>=0.10.2,<=0.11.0",
    "flash-attn>=2.8.1",
    "qwen-vl-utils",
]

sdk = [
    "litellm[proxy]",
    "aiosqlite",
    "asgiref",
]

tinker = [
    "tinker ; python_version >= '3.11'",
    "tinker-cookbook @ git+https://github.com/thinking-machines-lab/tinker-cookbook.git#egg=tinker-cookbook ; python_version >= '3.11'",
]

opentelemetry = [
    "opentelemetry-sdk",
]

smolagents = [
    "smolagents==1.20.0",
]

strands = [
    "strands-agents",
]

swe = [
    "docker",
    "kubernetes",
    "swebench",
]

web = [
    "selenium",
    "browsergym",
    "firecrawl",
    "gymnasium",
]

code-tools = [
    "e2b_code_interpreter",
    "together>=1.4",
]

[tool.uv.extra-build-dependencies]
flash-attn = [
    { requirement = "torch", match-runtime = true },
]

[tool.ruff]
line-length = 5000 # TODO: Reduce this to a more reasonable value

[tool.ruff.lint]
isort = {known-first-party = ["rllm"], known-third-party = ["verl"]}
# c.f. https://github.com/vllm-project/vllm/blob/ce8d6b75fc0586045df75ee1568a5b5f9957251b/pyproject.toml
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # isort
    "I",
    "G",
]
ignore = [
    # star imports
    "F405", "F403",
    # lambda expression assignment
    "E731",
    # Loop control variable not used within loop body
    "B007",
    # f-string format
    "UP032",
    # `.log()` statement uses f-string
    "G004",
    # equality check using x == True
    "E712", 
]

[tool.ruff.lint.per-file-ignores]
"rllm/rewards/code_utils/**" = ["ALL"]

[tool.setuptools.packages.find]
where   = ["."]
include = ["rllm", "rllm.*"]

[tool.setuptools.package-data]
rllm = ["trainer/config/*.yaml"]

[tool.mypy]
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
check_untyped_defs = true
follow_imports = "silent"
exclude = [
    "verl/.*",
    "rllm/rewards/code_utils/.*",
    "rllm/rewards/math_utils/.*",
    "examples/.*",
    "scripts/.*",
    "tests/.*",
    "docs/.*",
]